+++ packages/plugin/src/Bundles/Fields/FieldsBundle.php
@@ -31,6 +31,7 @@ use Solspace\Freeform\Fields\Implementations\Pro\SignatureField;
 use Solspace\Freeform\Fields\Implementations\Pro\TableField;
 use Solspace\Freeform\Fields\Implementations\Pro\WebsiteField;
 use Solspace\Freeform\Fields\Implementations\RadiosField;
+use Solspace\Freeform\Fields\Implementations\RangeField;
 use Solspace\Freeform\Fields\Implementations\TextareaField;
 use Solspace\Freeform\Fields\Implementations\TextField;
 use Solspace\Freeform\Library\Bundles\FeatureBundle;
@@ -68,6 +69,7 @@ class FieldsBundle extends FeatureBundle
             FileUploadField::class,
             FileDragAndDropField::class,
             NumberField::class,
+            RangeField::class,
 
             // Pro fields
             ConfirmationField::class,